refactor(shared): colocate IndexedDBStore with backup reads
Move IndexedDBStore and its tests to shared/lib so UI and background use
the same IndexedDB helper. PersistenceManager imports it from shared.

Rename readBackupFromIndexedDB to getVaultBackup and implement it via
IndexedDBStore; safeGetVaultBackup delegates to getVaultBackup.

Made-with: Cursor
